Fixed ls: cannot access '*_packagescan_*' error

+++ main.sh
@@ -29,9 +29,9 @@ fi
 
 
 #_ RENAME SYFT INDEXING RESULT FILES TO OLD, SO THAT THE VULNERABILITY REPORT WILL BE UP TO DATE _#
-for JSON_RESULT in $(ls *_packagescan_*)
+for JSON_RESULT in $(ls *_packagescan_* 2> /dev/null || true)
 do
-    mv ${JSON_RESULT} ${JSON_RESULT}.OLD
+    if [[ -f ${JSON_RESULT} ]]; then mv ${JSON_RESULT} ${JSON_RESULT}.OLD; fi
 done
